10000 updates (#17)
* updates signatures to match latest commit pkhex
updates tests to match new signature names
more to come, tests still broken

* still broken but heres some more progress.

* sorta solves some more rng issues, platinum is still broken

* fixes most things except gen 1/2 Mew and a few unowns in gen4

* - fixes unown issue

* I'm not entirely sure some of this is needed but it resolved the issues. undoes not great changes and applies a better fix. adds a fix for living dex.

* better handling for unown forms set by rng

* - adjustments for save block being Memory<byte> based

* random team generator

* random team generator

* random team generator adjustments

* random team generation adjustments

+++ PKHeX.Core.AutoMod/AutoMod/Regeneration/RegenUtil.cs
@@ -6,7 +6,7 @@ namespace PKHeX.Core.AutoMod
 {
     public static class RegenUtil
     {
-        public static bool GetTrainerInfo(IEnumerable<string> lines, int format, out ITrainerInfo tr)
+        public static bool GetTrainerInfo(IEnumerable<string> lines, byte format, out ITrainerInfo tr)
         {
             var sti = new SimpleTrainerInfo { Generation = format };
 
@@ -30,7 +30,7 @@ namespace PKHeX.Core.AutoMod
                         SID7 = sid;
                         break;
                     case "OTGender":
-                        sti.Gender = value is "Female" or "F" ? 1 : 0;
+                        sti.Gender = value is "Female" or "F" ? (byte)1 : (byte)0;
                         break;
                     default:
                         continue;
